+++ pyogrio/raw.py
@@ -39,6 +39,8 @@ def read(
     where=None,
     bbox=None,
     fids=None,
+    sql=None,
+    sql_dialect=None,
     return_fids=False,
 ):
     """Read OGR data source.
@@ -142,6 +144,8 @@ def read(
             where=where,
             bbox=bbox,
             fids=fids,
+            sql=sql,
+            sql_dialect=sql_dialect,
             return_fids=return_fids,
         )
     finally: